src/requirements.txt
kafka-python
requests
pandas
confluent-kafka
jupyterlab
